--- a/.env.example
+++ b/.env.example
@@ -36,6 +36,21 @@ LOG_FORMAT="text" # valores aceitos: text, json
 ENABLE_METRICS=false
 METRICS_PATH="/metrics"
 
+# Cache and Performance
+REDIS_URL=""
+CACHE_ENABLED=true
+CACHE_TTL_SUBJECTS=7200
+CACHE_TTL_CHARTS=14400
+CACHE_TTL_TRANSITS=1800
+
+# Rate Limiting
+RATE_LIMIT_ENABLED=true
+RATE_LIMIT_REQUESTS=60
+RATE_LIMIT_WINDOW=60
+
+# Compression
+ENABLE_COMPRESSION=true
+
 # Traefik / Portainer
 TRAEFIK_ENTRYPOINT="websecure"
 TRAEFIK_DOMAIN="api.exemplo.com"