… CLAUDE.md
… PERFORMANCE_OPTIMIZATION.md
… app/core/middleware.py
… app/main.py
… app/services/async_astrology_service.py
… pyproject.toml
… scripts/performance_benchmark.py
… tests/test_main.py
… tests/test_performance_simple.py